refactor: Migrate key code/modifier tests to their new modules
